Revert "refactor: keep "desgin" inconsistence across all linear model-based methods"
We can not use a numeric vector in which the first argument is -1, to
run pair-wisecomparison to the reference level , since the first
coefficient is intercept, and other coefficients are the difference
between the correponding level and reference level) whilesetting the
design as ~groups. If the first element of contrast is the same as the
reference level, the corresponding coef the base level is the
difference between base and reference level. Otherwise, both the
elements are the same as the reference level, a vector such as c(0, -1,
1) can be used to extract the difference between the third level and
second level.